requirements.txt
pandas==1.5.2
plotly-express==0.4.1
streamlit==1.15.2
plotly==5.11.0
Pillow==9.3.0
scikit-learn==1.1.3
scipy==1.9.3
toml==0.10.2
matplotlib==3.6.2
shap==0.41.0